refactor: test class 분리
