+++ src/test/java/step2/LottoNumbersTest.java
@@ -1,8 +1,6 @@
 package step2;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -11,29 +9,12 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class LottoTest {
-
-    @ParameterizedTest
-    @ValueSource(ints = {1, 45})
-    void createLottoNumberSuccessTest(int inputNumber) {
-        final LottoNumber lottoNumber = new LottoNumber(inputNumber);
-        final LottoNumber expected = new LottoNumber(inputNumber);
-        assertThat(lottoNumber).isEqualTo(expected);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {0, 46})
-    void createLottoNumberFailTest(int inputNumber) {
-        assertThatThrownBy(() -> new LottoNumber(inputNumber))
-                .isExactlyInstanceOf(RuntimeException.class)
-                .hasMessage("1~45 사이의 숫자만 생성 가능합니다.");
-    }
-
+public class LottoNumbersTest {
     @Test
     void createLottoNumbersDuplicateExceptionTest() {
         assertThatThrownBy(() -> new LottoNumbers(() -> 1))
                 .isExactlyInstanceOf(RuntimeException.class)
-                        .hasMessage("로또 숫자 생성 시간이 초과되었습니다.");
+                .hasMessage("로또 숫자 생성 시간이 초과되었습니다.");
     }
 
     @Test
@@ -65,11 +46,5 @@ public class LottoTest {
         assertThatThrownBy(() -> new LottoNumbers(Arrays.asList(1, 1, 1, 1, 1)))
                 .isExactlyInstanceOf(RuntimeException.class)
                 .hasMessage("로또 숫자는 6개여야 입니다.");
-    }
-
-    @Test
-    void name() {
-        new LottoNumbers(Arrays.asList(1, 2, 3, 4, 5, 6));
-        new WinningLotto(Arrays.asList(1, 2, 3, 4, 5, 6));
     }
 }
